refactor: refactor canvas module
