+++ src/modules/Canvas/components/Canvas.tsx
@@ -1,23 +1,23 @@
 import { Button, ButtonGroup, Container, useMediaQuery } from '@mui/material';
-import useCanvas from './useCanvas';
-import { useAppSelector } from '../../store/store';
+import useCanvas from '../hooks/useCanvas';
+import { useAppSelector } from '../../../store/store';
 import { toast } from 'react-toastify';
 import SaveIcon from '@mui/icons-material/Save';
 import { LoadingButton } from '@mui/lab';
 import { User } from 'firebase/auth';
-import useUploadImage from './useUploadImage';
+import useUploadImage from '../hooks/useUploadImage';
+import { useCanvasSize } from '../hooks/useCanvasSize';
 
 interface CanvasPropsRef {
   user: User | null | undefined;
 }
 
-function Canvas({ user }: CanvasPropsRef) {
+export function Canvas({ user }: CanvasPropsRef) {
   const editImage = useAppSelector((state) => state.editImage.editImageData);
   const tool = useAppSelector((state) => state.tool.toolValue);
   const toolsColor = useAppSelector((state) => state.toolColor.toolColorValue);
   const penSize = useAppSelector((state) => state.penSize.penSizeValue);
-  const isMobile = useMediaQuery('(max-width:570px)');
-  const isTablet = useMediaQuery('(max-width:780px)');
+  const { width, height } = useCanvasSize();
   const isSmallMobile = useMediaQuery('(max-width:430px)');
   const { canvasRef, clearCanvas, eventHandlers } = useCanvas(
     toolsColor,
@@ -32,26 +32,6 @@ function Canvas({ user }: CanvasPropsRef) {
   const handleUploadButtonClick = async () => {
     await uploadImage();
     toast.success('Image successfully uploaded!');
-  };
-
-  const getCanvasSize = () => {
-    if (isMobile) {
-      return {
-        width: 300,
-        height: 450,
-      };
-    }
-    if (isTablet) {
-      return {
-        width: 450,
-        height: 600,
-      };
-    }
-
-    return {
-      width: 600,
-      height: 750,
-    };
   };
 
   return (
@@ -70,8 +50,8 @@ function Canvas({ user }: CanvasPropsRef) {
         </ButtonGroup>
         <canvas
           ref={canvasRef}
-          height={getCanvasSize().height}
-          width={getCanvasSize().width}
+          height={height}
+          width={width}
           style={{
             alignSelf: 'center',
             background: '#fff',
@@ -88,5 +68,3 @@ function Canvas({ user }: CanvasPropsRef) {
     </>
   );
 }
-
-export default Canvas;
