+++ src/modules/Canvas/index.ts
@@ -1,0 +1,1 @@
+export { Canvas } from './components/Canvas';
