+++ src/modules/ImageGallery/index.ts
@@ -1,2 +1,3 @@
-export { ImageGallery } from './components/ImageGallery';
+export { ImageGallery } from './components/ImageGallery/ImageGallery';
+
 export { editImageReducer } from './store/editImageSlice';
