+++ src/pages/EditorPage/EditorPage.tsx
@@ -1,8 +1,8 @@
 import { useAuthState } from 'react-firebase-hooks/auth';
-import Canvas from '../../components/Canvas/Canvas';
 import ToolsPanel from '../../components/ToolsPanel/ToolsPanel';
 import { auth } from '../../api/config';
 import Spinner from '../../ui/Spinner/Spinner';
+import { Canvas } from '../../modules/Canvas';
 
 function EditorPage() {
   const [user, loading] = useAuthState(auth);
